+++ gempak/source/programs/dc/dcnmos/dcnmos.c
@@ -13,7 +13,7 @@
 
 void dcnmdc ( char *curtim, char *gemfil, char *prmfil, char *stntbl, 
                  int *iadstn, int *maxtim, int *nhours, int *txtflg, 
-                 int *iret, int, int, int, int );
+                 int *iret, size_t, size_t, size_t, size_t );
 
 int main ( int argc, char *argv[] )
 /************************************************************************
@@ -40,6 +40,7 @@ int main ( int argc, char *argv[] )
  * m.gamazaychikov/SAIC 07/05   Added parameters to CS of dc_gopt       *
  * H. Zeng/SAIC		08/05	Added parameters to CS of dc_gopt	*
  * L. Hinson/AWC        06/08   Added circflg parameter to dc_gopt      *
+ * SGuan-BHebbard/NCEP  06/22   Changed strlen in protos int->size_t    *
  ***********************************************************************/
 {
