+++ gempak/source/programs/gd/gdcross/gdxdta.c
@@ -6,22 +6,22 @@ void gdxdtaf  ( int *iflno, char *gdatim, char *gvcord, float *ystrt,
                float *rgx, float *rgy, int *nhxs, float *grid, 
                float *rlvl, float *xgrd,
                int *nvxs, char *parm, float *ybeg, float *yend, int *iret,
-               int gdatimsize, int gvcordsize, int gfuncsize, int timesize, 
-               int parmsize );
+               size_t gdatimsize, size_t gvcordsize, size_t gfuncsize, size_t timesize, 
+               size_t parmsize );
 
 void gdxdta  ( int *iflno, char *gdatim, char *gvcord, float *ystrt,
                float *ystop, char *gfunc, char *time, int *ivcord,
                float *rgx, float *rgy, int *nhxs, float *rlvl, float *xgrd,
                int *nvxs, char *parm, float *ybeg, float *yend, int *iret,
-               int gdatimsize, int gvcordsize, int gfuncsize, int timesize, 
-               int parmsize );
+               size_t gdatimsize, size_t gvcordsize, size_t gfuncsize, size_t timesize, 
+               size_t parmsize );
 
 void gdxdta  ( int *iflno, char *gdatim, char *gvcord, float *ystrt,
                float *ystop, char *gfunc, char *time, int *ivcord,
                float *rgx, float *rgy, int *nhxs, float *rlvl, float *xgrd,
                int *nvxs, char *parm, float *ybeg, float *yend, int *iret,
-               int gdatimsize, int gvcordsize, int gfuncsize, int timesize, 
-               int parmsize )
+               size_t gdatimsize, size_t gvcordsize, size_t gfuncsize, size_t timesize, 
+               size_t parmsize )
 /************************************************************************
  * gdxdta								*
  *									*
@@ -66,6 +66,7 @@ void gdxdta  ( int *iflno, char *gdatim, char *gvcord, float *ystrt,
  * Log:									*
  * S. Gilbert/NCEP       8/07   					*
  * S. Jacobs/NCEP	02/08	Set grid size as LLMXGD			*
+ * B. Hebbard/NCEP	08/21	Changed strlen in protos int->size_t    *
  ***********************************************************************/
 {
     int   kx, ky, kxy, ier;
